Train x encoder.

+++ src/train_x_encoder.py
@@ -36,8 +36,8 @@ if __name__ == '__main__':
     x_encoder_labels = np.load('data/processed/x_encoder_labels.npy')
 
     # Convert numpy arrays to PyTorch tensors
-    x_encoder_vectors = torch.from_numpy(x_encoder_vectors)
-    x_encoder_labels = torch.from_numpy(x_encoder_labels)
+    x_encoder_vectors = torch.from_numpy(x_encoder_vectors).long()
+    x_encoder_labels = torch.from_numpy(x_encoder_labels).long()
     x_encoder_labels = torch.unsqueeze(x_encoder_labels, 1)
 
     dataset = EncoderDataset(x_encoder_vectors, x_encoder_labels)
